requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8
pytz==2018.9
whitenoise==4.1.2
